divide resultLogs
