+++ src/front/interface.js
@@ -5,3 +5,4 @@
 export const Editor = require('./interface/aceditor');
 export const Panel = require('./interface/panel');
 export const Stdios = require('./interface/stdios');
+export const Results = require('./interface/results');
